+++ core/modules/modProductionHierarchy.class.php
@@ -64,7 +64,7 @@ class modProductionHierarchy extends DolibarrModules
 		$this->editor_url = 'https://www.wittkowski-it.de';
 
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-		$this->version = '1.0.0';
+		$this->version = '1.1.0';
 
 		// Key used in llx_const table to save module status enabled/disabled
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);